pyproject.toml
[tool.poetry]
name = "rmcraft-backend"
version = "0.1.0"
description = "Minimal FastAPI app with a /health endpoint"
authors = ["you <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.34.0", extras = ["standard"] }
google-adk = "^1.12.0"
google-generativeai = "^0.8.5"
dotenv = "^0.9.9"
sqlalchemy = "^2.0.43"
psycopg2-binary = "^2.9.10"
pydantic-settings = "^2.10.1"
chromadb = "*"
requests = "*"
beautifulsoup4 = "*"
jinja2 = "^3.1.6"
weasyprint = "^66.0"
cloudinary = "^1.44.1"
python-dotenv = "^1.1.1"
pypdf = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.0"
httpx = "*"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
